+++ Raytracer/src/camera.h
@@ -6,7 +6,7 @@
 class Camera
 {
 public:
-	virtual Ray ConstructRay(int pixelX, int pixelY) const = 0;
+	virtual Ray ConstructRay(float pixelX, float pixelY) const = 0;
 
 protected:
 	Camera(const Vector3& position, const Vector3& forward, int resolutionX, int resolutionY);
@@ -21,7 +21,7 @@ class OrthographicCamera : public Camera
 {
 public:
 	OrthographicCamera(const Vector3& position, const Vector3& forward, int resolutionX, int resolutionY, float size);
-	virtual Ray ConstructRay(int pixelX, int pixelY) const override;
+	virtual Ray ConstructRay(float pixelX, float pixelY) const override;
 
 private:
 	float m_size; // The camera can see size*2 units vertically and size*2*aspectRatio units horizontally
@@ -31,7 +31,7 @@ class PerspectiveCamera : public Camera
 {
 public:
 	PerspectiveCamera(const Vector3& position, const Vector3& forward, int resolutionX, int resolutionY, float fov = 45.0f);
-	virtual Ray ConstructRay(int pixelX, int pixelY) const override;
+	virtual Ray ConstructRay(float pixelX, float pixelY) const override;
 
 private:
 	float m_degreesFOV;
